fixed xml template resolver, minor refactor
