--- src/main/java/org/tdl/vireo/config/AppTemplateConfig.java
@@ -1,71 +1,0 @@
-package org.tdl.vireo.config;
-
-import static java.lang.Integer.MAX_VALUE;
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.thymeleaf.ThymeleafProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.ViewResolver;
-import org.thymeleaf.extras.springsecurity4.dialect.SpringSecurityDialect;
-import org.thymeleaf.spring4.SpringTemplateEngine;
-import org.thymeleaf.spring4.templateresolver.SpringResourceTemplateResolver;
-import org.thymeleaf.spring4.view.ThymeleafViewResolver;
-
-@Configuration
-@ConditionalOnClass({ SpringTemplateEngine.class })
-@EnableConfigurationProperties({ ThymeleafProperties.class })
-public class AppTemplateConfig implements ApplicationContextAware {
-
-    private ApplicationContext applicationContext;
-
-    @Autowired
-    private ThymeleafProperties properties;
-
-    public void setApplicationContext(ApplicationContext applicationContext) {
-        this.applicationContext = applicationContext;
-    }
-
-    @Bean
-    public ViewResolver viewResolver() {
-        ThymeleafViewResolver resolver = new ThymeleafViewResolver();
-        resolver.setOrder(MAX_VALUE);
-        resolver.setTemplateEngine(templateEngine());
-        resolver.setCharacterEncoding(UTF_8.name());
-        return resolver;
-    }
-
-    @Bean
-    public SpringTemplateEngine templateEngine() {
-        SpringTemplateEngine engine = new SpringTemplateEngine();
-        engine.addTemplateResolver(templateResolver());
-        engine.addTemplateResolver(xmlTemplateResolver());
-        engine.setEnableSpringELCompiler(true);
-        engine.addDialect(new SpringSecurityDialect());
-        return engine;
-    }
-
-    private SpringResourceTemplateResolver templateResolver() {
-        SpringResourceTemplateResolver resolver = new SpringResourceTemplateResolver();
-        resolver.setApplicationContext(applicationContext);
-        resolver.setPrefix(properties.getPrefix());
-        resolver.setSuffix(properties.getSuffix());
-        resolver.setTemplateMode(properties.getMode());
-        resolver.setCacheable(properties.isCache());
-        return resolver;
-    }
-
-    public SpringResourceTemplateResolver xmlTemplateResolver() {
-        SpringResourceTemplateResolver templateResolver = new SpringResourceTemplateResolver();
-        templateResolver.setPrefix("classpath:/formats/");
-        templateResolver.setSuffix(".xml");
-        templateResolver.setTemplateMode("XML");
-        return templateResolver;
-    }
-
-}
