+++ src/main/resources/application.properties
@@ -57,7 +57,7 @@ spring.thymeleaf.model: LEGACYHTML5
 logging.level.org.tdl: INFO
 logging.level.edu.tamu: INFO
 logging.level.org.hibernate: INFO
-logging.level.org.springframework: DEBUG
+logging.level.org.springframework: INFO
 logging.level.org.springframework.jndi: INFO
 logging.level.org.springframework.orm.jpa: INFO
